requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
gymnasium>=0.28.0
stable-baselines3>=2.0.0
sumolib>=1.15.0
traci>=1.15.0
streamlit>=1.22.0
plotly>=5.13.0
openrouter>=0.1.0
python-dotenv>=0.19.0 